% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/999_plot_palettes.R
\name{apply_palette}
\alias{apply_palette}
\title{Helper functions that deal with assignment of colour palettes for the overall plots}
\usage{
apply_palette(ggplot_object, palette_name)
}
\description{
Contains function \code{apply_palette}.
}
