% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/999_plot_kwargs.R
\name{assign_plot_kwargs}
\alias{assign_plot_kwargs}
\title{Helper functions that deal with assignment of plot_kwargs for plot}
\usage{
assign_plot_kwargs(dabest_effectsize_obj, plot_kwargs)
}
\description{
Contains function \code{assign_plot_kwargs}.
Function that handles plot_kwargs for final plotting
}
