% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_df_for_plots.R
\name{create_create_df_for_tufte}
\alias{create_create_df_for_tufte}
\title{Helper functions that generate dfs for plot}
\usage{
create_create_df_for_tufte(raw_data, enquo_x, enquo_y, proportional)
}
\description{
Contains functions \code{df_for_tufte}, \code{create_dfs_for_sankey}, \code{create_dfs_for_xaxis_redraw}, \code{create_dfs_for_proportion_bar} for generation of dfs.

Also includes \code{plot_slopegraph} (for plotting of slopegraph).
Function for creation of df for tuftelines plot
}
