% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_plot_components.R
\name{create_deltaplot_components}
\alias{create_deltaplot_components}
\title{Function for creation of list of TRUE/FALSE for delta plot components that will be built}
\usage{
create_deltaplot_components(
  proportional,
  is_paired,
  float_contrast,
  is_colour,
  delta2
)
}
\description{
Function for creation of list of TRUE/FALSE for delta plot components that will be built
}
