% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_df_for_plots.R
\name{create_dfs_for_sankey}
\alias{create_dfs_for_sankey}
\title{Function for creation of df for sankey plot}
\usage{
create_dfs_for_sankey(
  float_contrast = FALSE,
  raw_data,
  proportional_data,
  enquo_id_col,
  x_axis_raw,
  idx,
  scale_factor_sig = 0.8,
  bar_width = 0.15,
  gap,
  sankey = TRUE,
  flow = TRUE
)
}
\description{
Function for creation of df for sankey plot
}
