% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_df_for_plots.R
\name{create_dfs_for_xaxis_redraw}
\alias{create_dfs_for_xaxis_redraw}
\title{Function for creation of df for xaxis redraw for float_contrast FALSE plot}
\usage{
create_dfs_for_xaxis_redraw(idx)
}
\description{
Function for creation of df for xaxis redraw for float_contrast FALSE plot
}
