% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_plot_components.R
\name{create_rawplot_components}
\alias{create_rawplot_components}
\title{Helper functions that generates plot components.}
\usage{
create_rawplot_components(proportional, is_paired, float_contrast)
}
\description{
Contains functions \code{create_rawplot_components}, \code{create_deltaplot_components} and \code{create_violinplot_components}.
Function for creation of list of TRUE/FALSE for raw plot components that will be built
}
