% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_plot_components.R
\name{create_violinplot_components}
\alias{create_violinplot_components}
\title{Function for creation of list of values of the violin plot components that will be built}
\usage{
create_violinplot_components(
  boots,
  idx,
  float_contrast,
  delta_y_max,
  delta_y_min,
  flow = TRUE
)
}
\description{
Function for creation of list of values of the violin plot components that will be built
}
