% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_separate_plots.R
\name{create_xlabs_for_sankey}
\alias{create_xlabs_for_sankey}
\title{Function that creates xlabels for separated sankey diagrams}
\usage{
create_xlabs_for_sankey(idx, Ns)
}
\description{
Function that creates xlabels for separated sankey diagrams
}
