% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_plotter.R
\name{dabest_plot}
\alias{dabest_plot}
\title{Main plotting api}
\usage{
dabest_plot(dabest_effectsize_obj, float_contrast = TRUE, ...)
}
\description{
Contains function \code{dabest_plot}.

To be used after calculation of effect sizes with the various \code{effect_size} functions in _stat_tools/effsize.R.
}
