% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_data.R
\docType{data}
\name{deltadelta_data}
\alias{deltadelta_data}
\title{Data to produce a delta2 Dabest plot}
\format{
A data frame with 80 rows and 6 variables:
\describe{
\item{Dummy}{Dummy}
\item{Genotype}{Genotype}
\item{ID}{ID}
\item{Rep}{Rep}
\item{Treatment}{Which treatment method was used}
\item{Measurement}{Measurement}
}
}
\usage{
deltadelta_data
}
\description{
Contains 2 genotype groups and 2 treatment groups.
}
\examples{
data(deltadelta_data) # Lazy loading. Data becomes visible as soon as it is loaded
}
\keyword{datasets}
