% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_plot_geoms.R
\name{draw_group_halfviolin}
\alias{draw_group_halfviolin}
\title{Contains custom \if{html}{\out{<ggproto>}} geom_objects for plotting.}
\usage{
draw_group_halfviolin(data, panel_scales, coord)
}
\description{
Contains main geoms \code{geom_halfviolin}, \code{geom_bootci}, \code{geom_proportionbar}, \code{geom_sankeybar} and \code{geom_sankeyflow}.
}
