% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_bootstrap_tools.R
\name{effsize_boot}
\alias{effsize_boot}
\title{Obtains bootstraps as well as bca intervals}
\usage{
effsize_boot(data, effect_size_func, reps = 5000, paired = FALSE)
}
\description{
Contains functions \code{effsize_boot}, \code{bootstrap}, \code{bca} and \code{boot_weighted_row}.
}
