% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_plot_themes.R
\docType{data}
\name{float_contrast_theme}
\alias{float_contrast_theme}
\title{Contains \if{html}{\out{<ggplot>}} themes for float_contrast == TRUE | FALSE}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 6.
}
\usage{
float_contrast_theme
}
\description{
Contains themes \code{float_contrast_theme} and \code{non_float_contrast_theme}.
Theme for left-right graph
}
\keyword{datasets}
