% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_effsize.R
\name{mean_diff}
\alias{mean_diff}
\title{Main effectsize api}
\usage{
mean_diff(dabest_obj)
}
\description{
Contains functions \code{mean_diff}, \code{median_diff}, \code{cohens_d}, \code{hedges_g},
\code{cliffs_delta}, \code{cohens_h} and \code{hedges_correction}.

To be used after loading in data with \code{load()} function.
}
