% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_data.R
\docType{data}
\name{multigroup_data}
\alias{multigroup_data}
\title{Data to proudce multi-group Dabest plot}
\format{
A data frame with 240 rows and 5 variables:
\describe{
\item{Dummy}{Dummy}
\item{Gender}{Gender}
\item{ID}{ID}
\item{Group}{Which control group or test it is}
\item{Measurement}{Measurement}
}
}
\usage{
multigroup_data
}
\description{
Contains 3 control groups and 3 test groups.
}
\examples{
data(multigroup_data) # Lazy loading. Data becomes visible as soon as it is loaded
}
\keyword{internal}
