% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_data.R
\docType{data}
\name{my.data.proportional}
\alias{my.data.proportional}
\title{Data to proudce proportional or sankey Dabest plot (Can also be used multi-sankey)}
\format{
A data frame with 240 rows and 3 variables:
\describe{
\item{Group}{Which control group or test it is}
\item{Success}{1 or 0}
\item{ID}{ID}
}
}
\usage{
my.data.proportional
}
\description{
Contains 2 control groups and 4 test groups.
}
\examples{
data(my.data.proportional) # Lazy loading. Data becomes visible as soon as it is loaded
}
\keyword{datasets}
