% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_plot_themes.R
\docType{data}
\name{non_float_contrast_theme}
\alias{non_float_contrast_theme}
\title{Theme for top-down graph}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 7.
}
\usage{
non_float_contrast_theme
}
\description{
Theme for top-down graph
}
\keyword{datasets}
