% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_plot_api.R
\name{plot_raw}
\alias{plot_raw}
\title{Contains functions responsible for generation of raw_plot and delta_plot.}
\usage{
plot_raw(dabest_effectsize_obj, float_contrast, plot_kwargs)
}
\description{
Contains main plotting functions \code{plot_raw} and \code{plot_delta} for plotting of the rawdata and effectsize parts.
}
