% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_slopegraph.R
\name{plot_slopegraph}
\alias{plot_slopegraph}
\title{Functions that generate main_plots based on the main_plot_type of \code{create_rawplot_components}.}
\usage{
plot_slopegraph(dabest_effectsize_obj, plot_kwargs)
}
\description{
Contains function \code{plot_slopegraph}.
Function that plots slopegraph
}
