% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_plot_themes.R
\docType{data}
\name{remove_all_axes_theme}
\alias{remove_all_axes_theme}
\title{Theme for removal of all axes and labels}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 5.
}
\usage{
remove_all_axes_theme
}
\description{
Theme for removal of all axes and labels
}
\keyword{datasets}
