% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_separate_plots.R
\name{remove_last_ele_from_nested_list}
\alias{remove_last_ele_from_nested_list}
\title{Function that removes the last element from each subgroup within a list()}
\usage{
remove_last_ele_from_nested_list(nested_list)
}
\description{
Function that removes the last element from each subgroup within a list()
}
