% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_separate_plots.R
\name{separate_idx}
\alias{separate_idx}
\title{Contains functions responsible for separate plot functionalities via flow = FALSE.}
\usage{
separate_idx(idx, paired)
}
\description{
Contains \code{separate_idx}, \code{remove_last_ele_from_nested_list}, \code{create_xlabs_for_sankey} functions.
Separate idx function
}
