% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_formulas.R
\name{sigmoid}
\alias{sigmoid}
\title{Formulas used}
\usage{
sigmoid(x_from, factor, y_from, y_to, smooth = 5.5, n = 300)
}
\description{
Contains functions \code{sigmoid}, \code{flipped_sig}, \code{calculate_group_variance}, \code{calculate_weighted_delta}, \code{confinterval} and \code{var_w_df}.
}
