% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_data.R
\docType{data}
\name{twogroup_data}
\alias{twogroup_data}
\title{Data to produce the basic 2-group Dabest plot}
\format{
A data frame with 240 rows and 5 variables:
\describe{
\item{Dummy}{Dummy}
\item{Gender}{Gender}
\item{ID}{ID}
\item{Group}{Which control group or test it is}
\item{Measurement}{Measurement}
}
}
\usage{
twogroup_data
}
\description{
Contains 2 control groups and 4 test groups.
}
\examples{
data(twogroup_data) # Lazy loading. Data becomes visible as soon as it is loaded.
}
\keyword{datasets}
