% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_effsize_func.R
\name{effect_size}
\alias{effect_size}
\alias{mean_diff}
\alias{median_diff}
\alias{cohens_d}
\alias{hedges_g}
\alias{cliffs_delta}
\alias{cohens_h}
\title{Effect Size function for dabest_obj}
\usage{
mean_diff(dabest_obj)

median_diff(dabest_obj)

cohens_d(dabest_obj)

hedges_g(dabest_obj)

cliffs_delta(dabest_obj)

cohens_h(dabest_obj)
}
\arguments{
\item{dabest_obj}{dabest_obj created by loading in dataset along with other specified parameters with the \code{\link[=load]{load()}} function.}
}
\value{
A \code{dabest_effectsize_obj} list with 22 elements. The following are the elements contained within:
\itemize{
\item \code{raw_data} The dataset passed to \code{\link[=load]{load()}} that was cleaned and altered for plotting.
\item \code{idx} The list of control-test groupings as initially passed to \code{\link[=load]{load()}}.
\item \code{delta_x_labels} labels for the x-axis of the delta plot.
\item \code{delta_y_labels} labels for the y-axis of the delta plot.
\item \code{Ns} list of labels for x-axis of the raw plot.
\item \code{raw_y_labels} labels for the y-axis of the raw plot.
\item \code{is_paired} boolean value determining if it is a paired plot.
\item \code{is_colour} boolean value determining if there is a colour column for the plot.
\item \code{paired} paired ("sequential" or "baseline") as initially passed to \code{\link[=load]{load()}}.
\item \code{control_summary} value for plotting of control summary lines for float_contrast = \code{TRUE}.
\item \code{test_summary} value for plotting of control summary lines for float_contrast = \code{TRUE}.
\item \code{ylim} vector containing the y limits for the raw plot.
\item \code{enquo_x} quosure of x as initially passed to \code{\link[=load]{load()}}.
\item \code{enquo_y} quosure of y as initially passed to \code{\link[=load]{load()}}.
\item \code{enquo_id_col} quosure of id_col as initially passed to \code{\link[=load]{load()}}.
\item \code{enquo_colour} quosure of colour as initially passed to \code{\link[=load]{load()}}.
\item \code{proportional} boolean value as initially passed to \code{\link[=load]{load()}}.
\item \code{minimeta} boolean value as initially passed to \code{\link[=load]{load()}}.
\item \code{delta} boolean value as initially passed to \code{\link[=load]{load()}}.
\item \code{proportional_data} list of calculations related to the plotting of proportion plots.
\item \code{boot_result} list containing values related to the calculation of the effect sizes, bootstrapping and BCa correction.
\item \code{permtest_pvals} list containing values related to the calculations of permutation t tests and the corresponding p values,
and p values for different types of effect sizes and different statistical tests.
}
}
\description{
Calculates the effect size for each pairing of control and test group in \code{dabest_obj$idx}.
These five effect sizes \code{mean_diff}, \code{median_diff}, \code{cohens_d}, \code{hedges_g} and \code{cliffs_delta}
are used for most plot types.
}
\details{
The plot types listed under here are only able to use the following effect sizes.
\itemize{
\item Proportion plots offers only \code{mean_diff} and \code{cohens_h}
\item Mini-Meta Delta plots offers only \code{mean_diff}
}

The other plots are able to use all given basic effect sizes as listed in the Description.
}
\examples{
data <- data.frame(Group = c("Control1", "Control1", "Test1", "Test1"),
                   Measurement = c(100, 80, 50, 40)) 

dabest_obj <- load(data, x = Group, y = Measurement, idx = c("Control1", "Test1"))
dabest_obj.mean_diff <- mean_diff(dabest_obj)
dabest_obj.mean_diff

# Alternatively, you can pipe it using the `\%>\%` operator from the magrittr library like so
dabest_obj.mean_diff <- load(data, x = Group, y = Measurement, idx = c("Control1", "Test1")) \%>\% 
                          mean_diff()
dabest_obj.mean_diff

}
