% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_api.R
\name{load}
\alias{load}
\title{Main load api}
\usage{
load(
  data,
  x,
  y,
  idx = NULL,
  paired = NULL,
  id_col = NULL,
  ci = 95,
  colour = NULL,
  proportional = FALSE,
  minimeta = FALSE,
  delta2 = FALSE,
  experiment = NULL,
  experiment_label = NULL,
  x1_level = NULL
)
}
\arguments{
\item{data}{tidydataframe}

\item{x}{}

\item{y}{}

\item{idx}{}

\item{paired}{}

\item{id_col}{}

\item{ci}{}

\item{colour}{}

\item{proportional}{}

\item{minimeta}{}

\item{delta2}{}

\item{experiment}{}

\item{experiment_label}{}

\item{x1_level}{}
}
\value{
dabest_object
}
\description{
Main load api
}
