% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_api.R
\name{load}
\alias{load}
\title{Loading data with dabestr}
\usage{
load(
  data,
  x,
  y,
  idx = NULL,
  paired = NULL,
  id_col = NULL,
  ci = 95,
  colour = NULL,
  proportional = FALSE,
  minimeta = FALSE,
  delta2 = FALSE,
  experiment = NULL,
  experiment_label = NULL,
  x1_level = NULL
)
}
\arguments{
\item{data}{A tidy dataframe.}

\item{x}{Column name in the dateset that specifies the treatment group.}

\item{y}{Column name in the dataset that specifies the measurement values.}

\item{idx}{List of control-test groupings for which the
effect size will be computed for.}

\item{paired}{Paired ("sequential" or "baseline").}

\item{id_col}{Column name indicating the identity of the
datapoint if the data is paired. Must be supplied if paired is TRUE.}

\item{ci}{Double ranging from 0 to 100 which determines the bca_low and
bca_high confidence interval calculations}

\item{colour}{Column name indicating the column to implement colour aesthetics
in plotting functions.}

\item{proportional}{Boolean value determining whether the data provided is for
the plotting of proportional bars or sankey diagrams.}

\item{minimeta}{Boolean value determining if mini-meta analysis is conducted.}

\item{delta2}{Boolean value determining Whether delta-delta analysis for
2 by 2 experimental designs are conducted.}

\item{experiment}{Experiment column name for delta2 analysis.}

\item{experiment_label}{String specifying the experiment label that is used to distinguish
the experiment and the factors (being used in the plotting labels).}

\item{x1_level}{String setting the first factor level in
a 2 by 2 experimental design.}
}
\value{
Returns a dabest_object list with 18 elements. The following are the elements contained within:
\itemize{
\item \code{raw_data} The tidy dataframe that was passed into \code{\link[=load]{load()}}.
\item \code{proportional_data} List of calculations related to the plotting of proportion plots.
\item \code{enquo_x}  Quosure of x whose columns in the dataframe will be used to plot the x-axis.
\item \code{enquo_y}  Quosure of y whose columns in the dataframe will be used to plot the y-axis.
\item \code{enquo_id_col} Quosure of id_col whose columns in the dataframe will be used to
distinguish subjects in repeated measures.
\item \code{enquo_colour} Quosure of colour which will determine the color aesthetics of the plot.
\item \code{proportional} Same as above.
\item \code{minimeta}  Same as above.
\item \code{delta2}  Same as above.
\item \code{idx} Same as above.
\item \code{is_paired} Boolean value determining if it is a paired plot.
\item \code{is_colour} Boolean value determining if there is a colour column for the plot.
\item \code{paired} Same as above.
\item \code{ci} Numeric value which determines the confidence interval
specified for the effect size and bootstrap calculations.
\item \code{Ns} List of labels for x-axis of the raw plot.
\item \code{control_summary} Numeric value for plotting of control summary lines for float_contrast= TRUE.
\item \code{test_summary} Numeric value for plotting of test summary lines for float_contrast = TRUE.
}
\itemize{
\item \code{ylim} Vector containing the y limits for the raw plot.
}
}
\description{
Loads and converts a dataset into the dabestr format.
The output of this function is then used as an input for various procedural
functions within dabestr to create estimation plots.
}
\details{
The steps of producing an estimation plot with dabestr is as follows
\enumerate{
\item Loading in tidy dataset with \code{\link[=load]{load()}}.
\item Calculate effect sizes using \code{\link[=effect_size]{effect_size()}}.
\item Construct plot object using \code{\link[=dabest_plot]{dabest_plot()}}
}
}
\examples{
# Loading in of the dataset
data(twogroup_data)

# Creating a dabest object
dabest_obj <- load(data = twogroup_data, x = Group, y = Measurement, 
idx = c("Control1", "Group1"))

}
